"cards_parsing" is split into different files. Some imports are replaced by absolute imports. Module "cards_parsing" is replaced by package with sources where each source contains module with a source class. The "create_source_object" and "get_random_card" functions are moved to "game" module.
