+++ Imaginarium/__init__.py
@@ -1,4 +1,4 @@
-from . import cards_parsing
+from . import sources
 from . import exceptions
 from . import game
 from . import rules_setup
