+++ Imaginarium/sources/__init__.py
@@ -1,0 +1,2 @@
+from .BaseSource import BaseSource
+from .vk import Vk
